Move RemoteHostUtils to client package
